Changes codeVerifier after every use of promptAsync
